app_init done
