+++ pubspec.yaml
@@ -31,10 +31,6 @@ dependencies:
   flutter:
     sdk: flutter
 
-  animated_splash_screen: ^1.3.0
-  page_transition: ^2.0.9
-  lottie: ^2.3.0
-  http: ^0.13.5
 
 
   # The following adds the Cupertino Icons font to your application.
@@ -71,7 +67,8 @@ flutter:
 
   # To add assets to your application, add an assets section, like this:
   assets:
-  - assets/logo.png
+    - assets/images/
+  
      
 
   # An image asset can refer to one or more resolution-specific "variants", see
